[TOW-102] 리뷰 반영
* Feature 분리
